Having a module for each is overkill
