--- spec/support/crud/aggregatable.rb
@@ -1,28 +1,0 @@
-# Copyright (C) 2014-2015 MongoDB, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-module Mongo
-  module CRUD
-    module Aggregatable
-
-      private
-
-      def aggregate(collection)
-        collection.find.tap do |view|
-          view = view.batch_size(arguments['batchSize']) if arguments['batchSize']
-        end.aggregate(arguments['pipeline']).to_a
-      end
-    end
-  end
-end
